pyproject.toml
[tool.uv]
dev-dependencies = []
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple/"

[project]
name = "podcli"
version = "1.1.0"
description = "A complete podcast processing system with search, download, and transcription capabilities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "pydub>=0.25.1",
    "ffmpeg-python>=0.2.0",
    "loguru>=0.7.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "modelscope>=1.11.0",
    "gitpython>=3.1.43",
    "torch>=2.0.0",
    "transformers>=4.50.0",
    "qwen-asr>=0.0.6",
    "whisperx>=3.1.0",
    "click>=8.1.7",
    "feedparser>=6.0.10",
    "pyyaml>=6.0.1",
    "rich>=14.3.2",
    "keyboard>=0.13.5",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
podcli = "src.cli:cli"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
